+++ src/main/java/pikachu/parser/Parser.java
@@ -1,12 +1,12 @@
-package parser;
+package pikachu.parser;
 
-import task.Deadline;
-import task.Event;
-import task.Task;
-import task.ToDo;
-import task.TaskList;
+import pikachu.task.Deadline;
+import pikachu.task.Event;
+import pikachu.task.Task;
+import pikachu.task.ToDo;
+import pikachu.task.TaskList;
 
-import storage.Storage;
+import pikachu.storage.Storage;
 
 import java.io.IOException;
