+++ src/main/java/pikachu/task/Deadline.java
@@ -1,4 +1,4 @@
-package task;
+package pikachu.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
