+++ src/main/java/pikachu/task/Event.java
@@ -1,4 +1,4 @@
-package task;
+package pikachu.task;
 
 public class Event extends Task {
     protected String from;
